Altera tipo da pk do modelo candidato para big integer.

+++ receitas/models.py
@@ -38,7 +38,7 @@ class Candidato(models.Model):
     cargo = models.IntegerField(choices=CARGO_CHOICES)
     numero = models.IntegerField()
     partido = models.CharField(max_length=8)
-    id = models.IntegerField(primary_key=True)
+    id = models.BigIntegerField(primary_key=True)
 
     def __str__(self):
         return "{} - {}".format(self.nome, str(self.numero))
